data-engineering/coingecko_daily/requirements.txt
# python library dependencies
python-dotenv
pycoingecko #coingecko API wrapper
sqlalchemy
psycopg2-binary
pandas
mypy
